:package: user.entity.ts -> database.entity.ts - rename - split UserEntity to UserInfoEntity and UserEntity
